+++ app/components/textAngular/textAngular.js
@@ -2,93 +2,17 @@
 textAngular
 Author : Austin Anderson
 License : 2013 MIT
-Version 1.1.1
+Version 1.1.2
 
-Requirements: Angular 1.2.0, Angular ngSanitize module
-Optional Requirements: Bootstrap 3.0.0 and font-awesome for styling if you are using the default classes and icons.
-
-How to Use:
-
-1. Include textAngular.js in your project, alternatively grab all this code and throw it in your directives.js module file.
-2. In your HTML instantiate textAngular as an attribute or element, the only required attribute is the ng-model which is the variable to bind the content of the editor to, this is the standard ng-model allowing you to use validating filters on it.
-3. I reccommend using the following CSS in your stylesheet or a variant of to display the text box nicely:
-	
-.ta-editor{
-	min-height: 300px;
-	height: auto;
-	overflow: auto;
-	font-family: inherit;
-	font-size: 100%;
-}
-
-4. Have fun!
-
-Setting Options:
-
-Several options can be set through attributes on the HTML tag, these are;
-	- ta-toolbar: this should evaluate to an array of arrays. Each element is the name of one of the toolbar tools. The default is: [['h1', 'h2', 'h3', 'p', 'pre', 'bold', 'italics', 'ul', 'ol', 'redo', 'undo', 'clear'],['html', 'insertImage', 'insertLink']]
-	- ta-toolbar-class: this is the class to apply to the overall div of the toolbar, defaults to "btn-toolbar". Note that the class "ta-toolbar" is also added to the toolbar.
-	- ta-toolbar-group-class: this is the class to apply to the nested groups in the toolbar, a div with this class is created for each nested array in the ta-toolbar array and then the tool buttons are nested inside the group, defaults to "btn-group".
-	- ta-toolbar-button-class: this is the class to apply to each tool button in the toolbar, defaults to: "btn btn-default"
-	- ta-toolbar-active-button-class: this is the class to apply to each tool button in the toolbar if it's activeState function returns true ie when a tool function is applied to the selected text, defaults to: "active".
-	- ta-text-editor-class: this is the class to apply to the text editor <pre>, defaults to "form-control". Note that the classes: ta-editor and ta-text are also added.
-	- ta-html-editor-class: this is the class to apply to the html editor <div>, defaults to "form-control". Note that the classes: ta-editor and ta-html are also added.
-
-The defaults can be changed by altering/overwriting the variable: $rootScope.textAngularOpts which acts like global defaults for the classes and toolbar.
-The default value for this is: 
-	$rootScope.textAngularOpts = {
-		toolbar: [['h1', 'h2', 'h3', 'p', 'pre', 'bold', 'italics', 'ul', 'ol', 'redo', 'undo', 'clear'],['html', 'insertImage', 'insertLink']],
-		classes: {
-			toolbar: "btn-toolbar",
-			toolbarGroup: "btn-group",
-			toolbarButton: "btn btn-default",
-			toolbarButtonActive: "active",
-			textEditor: 'form-control',
-			htmlEditor: 'form-control'
-		}
-	}
-
-The toolbar buttons are defined in the object variable $rootScope.textAngularTools.
-The following is an example of how to add a button to make the selected text red:
-`
-$rootScope.textAngularTools.colourRed = {
-	display: "<button ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-square' style='color: red;'></i></button>",
-	action: function(){
-		this.$parent.wrapSelection('formatBlock', '<span style="color: red">');
-	},
-	activeState: function(){return false;} //This isn't required, and currently doesn't work reliably except for the html tag that doesn't rely on the cursor position.
-};
-//the following adds it to the toolbar to be displayed and used.
-$rootScope.textAngularOpts.toolbar = [['h1', 'h2', 'h3', 'p', 'pre', 'bold', 'colourRed', 'italics', 'ul', 'ol', 'redo', 'undo', 'clear'],['html', 'insertImage', 'insertLink']];
-`
-To explain how this works, when we create a button we create an isolated child scope of the textAngular scope and extend it with the values in the tools object, we then compile the HTML in the display value with the newly created scope.
-Note that the way any functions are called in the plugins the 'this' variable will allways point to the scope of the button ensuring that this.$parent will allways 
-Here's the code we run for every tool:
-
-`
-toolElement = angular.element($rootScope.textAngularTools[tool].display);
-toolElement.addClass(scope.classes.toolbarButton);
-groupElement.append($compile(toolElement)(angular.extend scope.$new(true), $rootScope.textAngularTools[tool]));
-`
+See README.md or https://github.com/fraywing/textAngular/wiki for requirements and use.
 */
 
+if(!window.console) console = {log: function() {}}; // fixes IE console undefined errors
 
 var textAngular = angular.module("textAngular", ['ngSanitize']); //This makes ngSanitize required
 
-textAngular.directive("textAngular", function($compile, $sce, $window, $document, $rootScope, $timeout) {
-	console.log("Thank you for using textAngular! http://www.textangular.com");
-	// Here we set up the global display defaults, make sure we don't overwrite any that the user may have already set.
-	$rootScope.textAngularOpts = angular.extend({
-		toolbar: [['h1', 'h2', 'h3', 'p', 'pre', 'bold', 'italics', 'ul', 'ol', 'redo', 'undo', 'clear'], ['html', 'insertImage', 'insertLink']],
-		classes: {
-			toolbar: "btn-toolbar",
-			toolbarGroup: "btn-group",
-			toolbarButton: "btn btn-default",
-			toolbarButtonActive: "active",
-			textEditor: 'form-control',
-			htmlEditor: 'form-control'
-		}
-	}, ($rootScope.textAngularOpts != null)? $rootScope.textAngularOpts : {});
+textAngular.directive("textAngular", ['$compile', '$window', '$document', '$rootScope', '$timeout', 'taFixChrome', function($compile, $window, $document, $rootScope, $timeout, taFixChrome) {
+	console.log("Thank you for using textAngular! http://www.textangular.com")
 	// deepExtend instead of angular.extend in order to allow easy customization of "display" for default buttons
 	// snatched from: http://stackoverflow.com/a/15311794/2966847
 	function deepExtend(destination, source) {
@@ -102,9 +26,27 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			}
 		}
 		return destination;
-	}
+	};
+	// Here we set up the global display defaults, make sure we don't overwrite any that the user may have already set.
+	$rootScope.textAngularOpts = deepExtend({
+		toolbar: [['h1', 'h2', 'h3', 'p', 'pre', 'quote'], ['bold', 'italics', 'underline', 'ul', 'ol', 'redo', 'undo', 'clear'], ['justifyLeft','justifyCenter','justifyRight'],['html', 'insertImage', 'insertLink', 'unlink']],
+		classes: {
+			focussed: "focussed",
+			toolbar: "btn-toolbar",
+			toolbarGroup: "btn-group",
+			toolbarButton: "btn btn-default",
+			toolbarButtonActive: "active",
+			textEditor: 'form-control',
+			htmlEditor: 'form-control'
+		}
+	}, ($rootScope.textAngularOpts != null)? $rootScope.textAngularOpts : {});
 	// Setup the default toolbar tools, this way allows the user to add new tools like plugins
-		$rootScope.textAngularTools = deepExtend({
+	var queryFormatBlockState = function(command){
+		command = command.toLowerCase();
+		var val = $document[0].queryCommandValue('formatBlock').toLowerCase();
+		return val === command || val === command;
+	}
+	$rootScope.textAngularTools = deepExtend({
 		html: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'>Toggle HTML</button>",
 			action: function() {
@@ -112,18 +54,14 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 				var ht, _this = this;
 				this.$parent.showHtml = !this.$parent.showHtml;
 				if (this.$parent.showHtml) { //Show the HTML view
-					ht = this.$parent.displayElements.text.html();
 					$timeout((function() { //defer until the element is visible
 						return _this.$parent.displayElements.html[0].focus(); //dereference the DOM object from the angular.element
 					}), 100);
 				} else { //Show the WYSIWYG view
-					ht = this.$parent.displayElements.html.html();
 					$timeout((function() { //defer until the element is visible
 						return _this.$parent.displayElements.text[0].focus(); //dereference the DOM object from the angular.element
 					}), 100);
 				}
-				// don't need to wrap in $apply as this is called within the ng-click which is in an $apply anyway
-				this.$parent.compileHtml(ht);
 				this.active = this.$parent.showHtml;
 			}
 		},
@@ -131,49 +69,57 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'>H1</button>",
 			action: function() {
 				return this.$parent.wrapSelection("formatBlock", "<H1>");
-			}
+			},
+			activeState: function() { return queryFormatBlockState('h1'); }
 		},
 		h2: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'>H2</button>",
 			action: function() {
 				return this.$parent.wrapSelection("formatBlock", "<H2>");
-			}
+			},
+			activeState: function() { return queryFormatBlockState('h2'); }
 		},
 		h3: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'>H3</button>",
 			action: function() {
 				return this.$parent.wrapSelection("formatBlock", "<H3>");
-			}
+			},
+			activeState: function() { return queryFormatBlockState('h3'); }
 		},
 		p: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'>P</button>",
 			action: function() {
 				return this.$parent.wrapSelection("formatBlock", "<P>");
-			}
+			},
+			activeState: function() { return queryFormatBlockState('p'); }
 		},
 		pre: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'>pre</button>",
 			action: function() {
 				return this.$parent.wrapSelection("formatBlock", "<PRE>");
-			}
+			},
+			activeState: function() { return queryFormatBlockState('pre'); }
 		},
 		ul: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-list-ul'></i></button>",
 			action: function() {
 				return this.$parent.wrapSelection("insertUnorderedList", null);
-			}
+			},
+			activeState: function() { return $document[0].queryCommandState('insertUnorderedList'); }
 		},
 		ol: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-list-ol'></i></button>",
 			action: function() {
 				return this.$parent.wrapSelection("insertOrderedList", null);
-			}
+			},
+			activeState: function() { return $document[0].queryCommandState('insertOrderedList'); }
 		},
 		quote: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-quote-right'></i></button>",
 			action: function() {
 				return this.$parent.wrapSelection("formatBlock", "<BLOCKQUOTE>");
-			}
+			},
+			activeState: function() { return queryFormatBlockState('blockquote'); }
 		},
 		undo: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-undo'></i></button>",
@@ -232,10 +178,19 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 				return $document[0].queryCommandState('italic');
 			}
 		},
+		underline: {
+			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-underline'></i></button>",
+			action: function() {
+				return this.$parent.wrapSelection("underline", null);
+			},
+			activeState: function() {
+				return $document[0].queryCommandState('underline');
+			}
+		},
 		clear: {
 			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-ban'></i></button>",
 			action: function() {
-				return this.$parent.wrapSelection("FormatBlock", "<div>");
+				return this.$parent.wrapSelection("removeFormat", null);
 			}
 		},
 		insertImage: {
@@ -249,7 +204,7 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			}
 		},
 		insertLink: {
-			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-chain'></i></button>",
+			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-link'></i></button>",
 			action: function() {
 				var urlLink;
 				urlLink = prompt("Please enter an URL to insert", 'http://');
@@ -257,14 +212,15 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 					return this.$parent.wrapSelection('createLink', urlLink);
 				}
 			}
+		},
+		unlink: {
+			display: "<button type='button' ng-click='action()' ng-class='displayActiveToolClass(active)'><i class='fa fa-unlink'></i></button>",
+			action: function() {
+				return this.$parent.wrapSelection('unlink', null);
+			}
 		}
 	}, ($rootScope.textAngularTools != null)? $rootScope.textAngularTools : {});
-	
-	// ngSanitize is a requirement for the module so this shouldn't cause any trouble
-	var sanitizationWrapper = function(html) {
-		return $sce.trustAsHtml(html);
-	};
-	
+		
 	return {
 		require: 'ngModel',
 		scope: {},
@@ -273,42 +229,25 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			var group, groupElement, keydown, keyup, tool, toolElement; //all these vars should not be accessable outside this directive
 			// get the settings from the defaults and add our specific functions that need to be on the scope
 			angular.extend(scope, $rootScope.textAngularOpts, {
-				// This must be called within a $apply or the ngModel value will not be updated correctly
-				compileHtml: function(html) {
-					// this refers to the scope
-					var compHtml = angular.element("<div>").append(html).html().replace(/(class="(.*?)")|(class='(.*?)')/g, "").replace(/&lt;/g, "<").replace(/&gt;/g, ">").replace(/style=("|')(.*?)("|')/g, "");
-					if (scope.showHtml === "load") {
-						scope.text = sanitizationWrapper(compHtml);
-						scope.html = sanitizationWrapper(compHtml.replace(/</g, "&lt;"));
-						scope.showHtml = scope.showHtmlDefault || false;
-					} else if (scope.showHtml) { // update the raw HTML view
-						scope.text = sanitizationWrapper(compHtml);
-					} else { // update the WYSIWYG view
-						scope.html = sanitizationWrapper(compHtml.replace(/</g, "&lt;"));
-					}
-					ngModel.$setViewValue(compHtml);
-				},
 				// wraps the selection in the provided tag / execCommand function.
-				wrapSelection: function(command, opt, updateDisplay) {
-					// the default value for updateDisplay is true
-					if (updateDisplay == null) updateDisplay = true;
+				wrapSelection: function(command, opt) {
 					document.execCommand(command, false, opt);
+					// strip out the chrome specific rubbish that gets put in when using lists
+					if(command === 'insertUnorderedList' || command === 'insertOrderedList') taFixChrome(scope.displayElements.text);
 					// refocus on the shown display element, this fixes a display bug when using :focus styles to outline the box. You still have focus on the text/html input it just doesn't show up
 					if (scope.showHtml)
-						scope.displayElements.text[0].focus();
-					else
 						scope.displayElements.html[0].focus();
+					else
+						scope.displayElements.text[0].focus();
 					// note that wrapSelection is called via ng-click in the tool plugins so we are already within a $apply
-					if (updateDisplay) scope.updateDisplay();
-				},
-				// due to restritions on compileHtml this must also be called in a scope.$apply - this is really a convenience function for compileHtml anyway.
-				updateDisplay: function() {
-					scope.compileHtml((!scope.showHtml)? scope.displayElements.html.html() : scope.displayElements.text.html());
+					scope.updateSelectedStyles();
+					if (!scope.showHtml) scope.updateTaBindtext(); // only update if NOT in html mode
 				},
 				showHtml: false
 			});
 			// setup the options from the optional attributes
 			if (!!attrs.taToolbar)					scope.toolbar = scope.$eval(attrs.taToolbar);
+			if (!!attrs.taFocussedClass)			scope.classes.focussed = scope.$eval(attrs.taFocussedClass);
 			if (!!attrs.taToolbarClass)				scope.classes.toolbar = attrs.taToolbarClass;
 			if (!!attrs.taToolbarGroupClass)		scope.classes.toolbarGroup = attrs.taToolbarGroupClass;
 			if (!!attrs.taToolbarButtonClass)		scope.classes.toolbarButton = attrs.taToolbarButtonClass;
@@ -319,13 +258,19 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			// Setup the HTML elements as variable references for use later
 			scope.displayElements = {
 				toolbar: angular.element("<div></div>"),
-				text: angular.element("<pre contentEditable='true' ng-show='showHtml' ng-bind-html='html' ></pre>"),
-				html: angular.element("<div contentEditable='true' ng-hide='showHtml' ng-bind-html='text' ></div>")
+				forminput: angular.element("<input type='hidden' style='display: none;'>"),
+				html: angular.element("<textarea ng-show='showHtml' ta-bind='html' ng-model='html' ></textarea>"),
+				text: angular.element("<div contentEditable='true' ng-hide='showHtml' ta-bind='text' ng-model='text' ></div>")
 			};
 			// add the main elements to the origional element
 			element.append(scope.displayElements.toolbar);
 			element.append(scope.displayElements.text);
 			element.append(scope.displayElements.html);
+			
+			if(!!attrs.name){
+				scope.displayElements.forminput.attr('name', attrs.name);
+				element.append(scope.displayElements.forminput);
+			}
 			
 			// compile the scope with the text and html elements only - if we do this with the main element it causes a compile loop
 			$compile(scope.displayElements.text)(scope);
@@ -336,6 +281,21 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			scope.displayElements.toolbar.addClass("ta-toolbar " + scope.classes.toolbar);
 			scope.displayElements.text.addClass("ta-text ta-editor " + scope.classes.textEditor);
 			scope.displayElements.html.addClass("ta-html ta-editor " + scope.classes.textEditor);
+			
+			// note that focusout > focusin is called everytime we click a button
+			element.on('focusin', function(){ // cascades to displayElements.text and displayElements.html automatically.
+				element.addClass(scope.classes.focussed);
+				$timeout(function(){ element.triggerHandler('focus'); }, 0); // to prevent multiple apply error defer to next seems to work.
+			});
+			element.on('focusout', function(){
+				$timeout(function(){
+					// if we have NOT focussed again on the text etc then fire the blur events
+					if(!($document[0].activeElement === scope.displayElements.html[0]) && !($document[0].activeElement === scope.displayElements.text[0])){
+						element.removeClass(scope.classes.focussed);
+						$timeout(function(){ element.triggerHandler('blur'); }, 0); // to prevent multiple apply error defer to next seems to work.
+					}
+				}, 0);
+			});
 			
 			scope.tools = {}; // Keep a reference for updating the active states later
 			// create the tools in the toolbar
@@ -350,7 +310,13 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 					toolElement = angular.element($rootScope.textAngularTools[tool].display);
 					toolElement.addClass(scope.classes.toolbarButton);
 					toolElement.attr('unselectable', 'on'); // important to not take focus from the main text/html entry
+					toolElement.attr('ng-disabled', 'showHtml()');
 					var childScope = angular.extend(scope.$new(true), $rootScope.textAngularTools[tool], { // add the tool specific functions
+						name: tool,
+						showHtml: function(){
+							if(this.name !== 'html') return this.$parent.showHtml;
+							return false;
+						},
 						displayActiveToolClass: function(active){
 							return (active)? this.$parent.classes.toolbarButtonActive : '';
 						}
@@ -363,14 +329,26 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			
 			// changes to the model variable from outside the html/text inputs
 			ngModel.$render = function() {
+				scope.displayElements.forminput.val(ngModel.$viewValue);
 				if(ngModel.$viewValue === undefined) return;
 				// if the editors aren't focused they need to be updated, otherwise they are doing the updating
 				if (!($document[0].activeElement === scope.displayElements.html[0]) && !($document[0].activeElement === scope.displayElements.text[0])) {
 					var val = ngModel.$viewValue || ''; // in case model is null
-					scope.text = sanitizationWrapper(val);
-					scope.html = sanitizationWrapper(val.replace(/</g, "&lt;"));
+					scope.text = val;
+					scope.html = val;
 				}
 			};
+			
+			scope.$watch('text', function(newValue, oldValue){
+				scope.html = newValue;
+				ngModel.$setViewValue(newValue);
+				scope.displayElements.forminput.val(newValue);
+			});
+			scope.$watch('html', function(newValue, oldValue){
+				scope.text = newValue;
+				ngModel.$setViewValue(newValue);
+				scope.displayElements.forminput.val(newValue);
+			});
 			
 			// the following is for applying the active states to the tools that support it
 			scope.bUpdateSelectedStyles = false;
@@ -399,7 +377,6 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			// stop updating on key up and update the display/model
 			keyup = function(e) {
 				scope.bUpdateSelectedStyles = false;
-				scope.$apply(scope.updateDisplay);
 			};
 			scope.displayElements.html.on('keyup', keyup);
 			scope.displayElements.text.on('keyup', keyup);
@@ -413,4 +390,83 @@ textAngular.directive("textAngular", function($compile, $sce, $window, $document
 			scope.displayElements.text.on('mouseup', mouseup);
 		}
 	};
+}]).directive('taBind', ['$sanitize', '$document', 'taFixChrome', function($sanitize, $document, taFixChrome){
+	// Uses for this are textarea or input with ng-model and ta-bind='text' OR any non-form element with contenteditable="contenteditable" ta-bind="html|text" ng-model
+	return {
+		require: 'ngModel',
+		scope: {'taBind': '@'},
+		link: function(scope, element, attrs, ngModel){
+			var isContentEditable = element[0].tagName.toLowerCase() !== 'textarea' && element[0].tagName.toLowerCase() !== 'input' && element.attr('contenteditable') !== undefined;
+			// in here we are undoing the converts used elsewhere to prevent the < > and & being displayed when they shouldn't in the code.
+			var compileHtml = function(){
+				var result = taFixChrome(angular.element("<div>").append(element.html())).html();
+				if(scope.taBind === 'html' && isContentEditable) result = result.replace(/&lt;/g, "<").replace(/&gt;/g, ">").replace(/&amp;/g, '&');
+				return result;
+			};
+			
+			scope.$parent['updateTaBind' + scope.taBind] = function(){//used for updating when inserting wrapped elements
+				var compHtml = compileHtml();
+				var tempParsers = ngModel.$parsers;
+				ngModel.$parsers = []; // temp disable of the parsers
+				ngModel.$oldViewValue = compHtml;
+				ngModel.$setViewValue(compHtml);
+				ngModel.$parsers = tempParsers;
+			};
+			
+			//this code is used to update the models when data is entered/deleted
+			if(isContentEditable){
+				element.on('keyup', function(e){
+					ngModel.$setViewValue(compileHtml());
+				});
+			}
+			
+			ngModel.$parsers.push(function(value){
+				// all the code here takes the information from the above keyup function or any other time that the viewValue is updated and parses it for storage in the ngModel
+				if(ngModel.$oldViewValue === undefined) ngModel.$oldViewValue = value;
+				try{
+					$sanitize(value); // this is what runs when ng-bind-html is used on the variable
+				}catch(e){
+					return ngModel.$oldViewValue; //prevents the errors occuring when we are typing in html code
+				}
+				ngModel.$oldViewValue = value;
+				return value;
+			});
+			
+			// changes to the model variable from outside the html/text inputs
+			ngModel.$render = function() {
+				if(ngModel.$viewValue === undefined) return;
+				// if the editor isn't focused it needs to be updated, otherwise it's receiving user input
+				if ($document[0].activeElement !== element[0]) {
+					var val = ngModel.$viewValue || ''; // in case model is null
+					ngModel.$oldViewValue = val;
+					if(scope.taBind === 'text'){ //WYSIWYG Mode
+						element.html(val);
+						element.find('a').on('click', function(e){
+							e.preventDefault();
+							return false;
+						});
+					}else if(isContentEditable || (element[0].tagName.toLowerCase() !== 'textarea' && element[0].tagName.toLowerCase() !== 'input')) // make sure the end user can SEE the html code.
+						element.html(val.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, '&gt;'));
+					else element.val(val); // only for input and textarea inputs
+				}else if(!isContentEditable) element.val(val); // only for input and textarea inputs
+			};
+		}
+	};
+}]).factory('taFixChrome', function(){
+	// get whaterever rubbish is inserted in chrome
+	var taFixChrome = function($html){ // should be an angular.element object, returns object for chaining convenience
+		// fix the chrome trash that gets inserted sometimes
+		var spans = angular.element($html).find('span'); // default wrapper is a span so find all of them
+		for(var s = 0; s < spans.length; s++){
+			var span = angular.element(spans[s]);
+			if(span.attr('style') && span.attr('style').match(/line-height: 1.428571429;|color: inherit; line-height: 1.1;/i)){ // chrome specific string that gets inserted into the style attribute, other parts may vary. Second part is specific ONLY to hitting backspace in Headers
+				if(span.next().length > 0 && span.next()[0].tagName === 'BR') span.next().remove()
+				span.replaceWith(span.html());
+			}
+		}
+		var result = $html.html().replace(/style="[^"]*?(line-height: 1.428571429;|color: inherit; line-height: 1.1;)[^"]*"/ig, ''); // regex to replace ONLY offending styles - these can be inserted into various other tags on delete
+		$html.html(result);
+		return $html;
+	};
+	return taFixChrome;
 });