+++ app/components/textAngular/changelog.md
@@ -1,5 +1,34 @@
 ###Changelog
 
+2013-12-11 v1.1.2
+
+- Updated to work correctly with IE (console.log bug)
+
+2013-12-11 v1.1.2-pre3
+
+- Added support for .focussed class and ng-focus to allow dynamic styling on focus events. #47
+- Updates to fix Angular.JS breaking with parameter renaming minification. #49
+- Minor bug fix to disable links from being 'clickable' in the editor.
+- Updated the default toolbar to include ALL default tools.
+- Update the tools to use activeState better.
+- Small update to allow use of ta-bind outside of textAngular.
+- Changed the raw html view to use a text-area for better compatability.
+
+2013-12-09 v1.1.2-pre2
+
+- Added input for form submission. #43
+- Slight restructure and update into of /demo.
+- Moved a lot of the README.md to the online Wiki. #34
+- Changed pre-release tag names to -preX as we aren't really doing alpha - beta - RC format.
+
+2013-12-05 v1.1.2-alpha (v1.1.2-pre1)
+
+- Added bundled demo pages.
+- Fixed Escaping of < and > #30
+- Fixed stripping of style and class attributes and other parsing issues whilst maintaining the chrome fixes. #35 #30 #5
+- Fixed two-way-binding not working #38
+- Updated Readme.md and consolidated the readme out of the textAngular.js file.
+
 2013-12-2 v1.1.1
 
 - Fixed buttons still submitting form. #29